Optimizations of sequential and deep voxel analysis

+++ rsa3d/utils/Utils.cpp
@@ -32,16 +32,16 @@ bool increment(int* in, int inlength, int max){
 }
 
 
-int position2i(const double* da, int dalength, double size, double dx, int n){
-	int result = 0;
+size_t position2i(const double* da, unsigned short int dalength, double size, double dx, size_t n){
+	size_t result = 0;
 	int ix;
 	for(int i=dalength-1; i>=0; i--){
 		if (da[i]<0)
-			ix = (int)((da[i] + size)/dx);
+			ix = static_cast<int>((da[i] + size) / dx);
 		else if (da[i]>=size)
-			ix = (int)((da[i] - size)/dx);
+			ix = static_cast<int>((da[i] - size) / dx);
 		else
-			ix = (int)(da[i]/dx);
+			ix = static_cast<int>(da[i] / dx);
 
 		if ( (ix+1)*dx == da[i])
 			ix++;
@@ -61,11 +61,11 @@ void i2position(double* da, int dalength, int index, double dx, int n){
 void coordinates(int* result, const double* da, int dalength, double size, double dx, int n){
 	for(int i=dalength-1; i>=0; i--){
 		if (da[i]<0)
-			result[i] = (int)((da[i] + size)/dx);
+			result[i] = static_cast<int>((da[i] + size) / dx);
 		else if (da[i]>=size)
-			result[i] = (int)((da[i] - size)/dx);
+			result[i] = static_cast<int>((da[i] - size) / dx);
 		else
-			result[i] = (int)(da[i]/dx);
+			result[i] = static_cast<int>(da[i] / dx);
 	}
 }
